# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicBox API Documentation",
        "description": "API Documentation for MusicBox",
        "contact": {
            "email": "your-email@domain.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Get list of songs",
                "operationId": "7cff74aca8adedf68f8a9248996cdc0d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "songs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Song Title"
                                                    },
                                                    "duration": {
                                                        "type": "integer",
                                                        "example": 180
                                                    },
                                                    "album_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Songs Fetched Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Create a new song",
                "operationId": "a28be4ced5c180620b197c9144c70376",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "duration",
                                    "album_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Song"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 180
                                    },
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Song created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "song": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Song Created Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/songs/{id}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Get song by ID with album and artist",
                "operationId": "485f590f1416afbb2a0aab6b12305004",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Song ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "song": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Song Title"
                                                },
                                                "duration": {
                                                    "type": "integer",
                                                    "example": 180
                                                },
                                                "album": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Album Title"
                                                        },
                                                        "artist": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Artist Name"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Song Fetched Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Song not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Songs"
                ],
                "summary": "Update existing song",
                "operationId": "caa5ed09d1ac8aeae370bbe30ae55004",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Song ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Song Title"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 200
                                    },
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Song updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "song": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Song Updated Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Song not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Songs"
                ],
                "summary": "Delete a song",
                "operationId": "ede008b9ebf14f39a6479db6065feaea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Song ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Song deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Song Deleted Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Song not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/songs/search": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Search songs by title or artist",
                "operationId": "486ea09f311b6528edc8cbe1d46f21a8",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Search by song title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "artist",
                        "in": "query",
                        "description": "Search by artist name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Song Title"
                                                    },
                                                    "duration": {
                                                        "type": "integer",
                                                        "example": 180
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No songs found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}